+++ package.json
@@ -20,6 +20,7 @@
     "typescript": "^2.7.1"
   },
   "devDependencies": {
+    "@types/express": "^4.11.1",
     "@types/next": "^2.4.7",
     "@types/react": "^16.0.36",
     "lint-staged": "^7.1.3",
@@ -28,7 +29,13 @@
     "ts-node": "^6.0.5"
   },
   "lint-staged": {
-    "*.{js,ts,tsx}": ["yarn prettier", "git add"],
-    "*.{json,css,md}": ["yarn prettier", "git add"]
+    "*.{js,ts,tsx}": [
+      "yarn prettier",
+      "git add"
+    ],
+    "*.{json,css,md}": [
+      "yarn prettier",
+      "git add"
+    ]
   }
 }
